requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.134
botocore==1.34.134
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-storages==1.14.3
djangorestframework==3.15.2
idna==3.7
jmespath==1.0.1
kombu==5.3.7
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg==3.2.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
